requirements.txt
# ── Web framework ──────────────────────────────────────────────────────────────
fastapi==0.115.6
uvicorn[standard]==0.32.1
sse-starlette==2.1.3
python-multipart==0.0.20

# ── Database / async ───────────────────────────────────────────────────────────
sqlalchemy==2.0.36
aiosqlite==0.20.0
pydantic==2.10.3
pydantic-settings==2.6.1

# ── LangChain ecosystem ────────────────────────────────────────────────────────
langchain==0.3.13
langchain-anthropic==0.3.3
langchain-community==0.3.13
langchain-core==0.3.30
langgraph==0.2.60
langchain-chroma==0.1.4
chromadb==0.5.23

# ── AI / Search APIs ───────────────────────────────────────────────────────────
anthropic>=0.41.0,<1
tavily-python==0.5.0

# ── Market data (Massive.com / Polygon.io — official API, no scraping) ─────────
polygon-api-client>=1.12.0
finnhub-python>=2.4.0

# ── Data / utilities ───────────────────────────────────────────────────────────
pandas>=1.5.0,<3
numpy>=1.22.4,<2
scipy>=1.10.0
apscheduler==3.10.4
httpx>=0.24.0,<1
aiohttp>=3.8.3,<4
beautifulsoup4==4.12.3
requests>=2.28.0,<3
python-dotenv==1.0.1
